Add createAlbum call
To create an album via the api we need to create two enums which represent the Privacy and Layout settings for the album.
We also have a new package (models.post) which will represent the objects which can be posted to the api, since they differ from the get objects (model.Album vs. model.post.Album)

refs #3